requirements.txt
numpy~=1.23.5
opencv-python~=4.7.0.72
mediapipe~=0.9.2.1
tensorflow~=2.12.0
pyserial~=3.5
keyboard~=0.13.5